Transition to new types
Closes https://github.com/PistonDevelopers/graphics/issues/918
