+++ src/rect.rs
@@ -1,115 +1,10 @@
 //! Contains types used in this library
 use std::convert::From;
-use std::ops::{ Add, Mul, Sub };
+
+use Size;
+use Point;
 
 pub use math::{ self, Matrix2d, Scalar, Vec2d };
-
-/// The size of a shape.
-#[derive(Clone, Copy, Debug)]
-pub struct Size {
-    /// The horizontal length of the shape (width).
-    pub w: Scalar,
-    /// The vertical length of the shape (height).
-    pub h: Scalar,
-}
-
-impl From<Vec2d> for Size {
-    fn from(v: Vec2d) -> Size {
-        Size { w: v[0], h: v[1] }
-    }
-}
-
-impl From<(Scalar, Scalar)> for Size {
-    fn from((w, h): (Scalar, Scalar)) -> Size {
-        Size { w: w, h: h }
-    }
-}
-
-impl Size {
-    /// Convert size to a vector.
-    pub fn to_array(self) -> Vec2d {
-        [self.w, self.h]
-    }
-}
-
-impl<T: Into<Size>> Mul<T> for Size {
-    type Output = Size;
-
-    fn mul(self, v: T) -> Size {
-        let v: Size = v.into();
-        Size { w: self.w * v.w, h: self.h * v.h }
-    }
-}
-
-impl Mul<Scalar> for Size {
-    type Output = Size;
-
-    fn mul(self, s: Scalar) -> Size {
-        Size { w: self.w * s, h: self.h * s }
-    }
-}
-
-/// A point in the Cartesian plane.
-#[derive(Clone, Copy, Debug)]
-pub struct Point {
-    /// The x coordinate.
-    pub x: Scalar,
-    /// The y coordinate.
-    pub y: Scalar,
-}
-
-impl Add<Scalar> for Point {
-    type Output = Point;
-
-    fn add(self, s: Scalar) -> Point {
-        Point { x: self.x + s, y: self.y + s }
-    }
-}
-
-impl<T: Into<Point>> Add<T> for Point {
-    type Output = Point;
-
-    fn add(self, v: T) -> Point {
-        let v: Point = v.into();
-        Point { x: self.x + v.x, y: self.y + v.y }
-    }
-}
-
-impl From<Vec2d> for Point {
-    fn from(v: Vec2d) -> Point {
-        Point { x: v[0], y: v[1] }
-    }
-}
-
-impl From<(Scalar, Scalar)> for Point {
-    fn from((x, y): (Scalar, Scalar)) -> Point {
-        Point { x: x, y: y }
-    }
-}
-
-impl Sub<Scalar> for Point {
-    type Output = Point;
-
-    fn sub(self, s: Scalar) -> Point {
-        Point { x: self.x - s, y: self.y - s }
-    }
-}
-
-impl<T: Into<Point>> Sub<T> for Point {
-    type Output = Point;
-
-    fn sub(self, v: T) -> Point {
-        let v = v.into();
-        Point { x: self.x - v.x, y: self.y - v.y }
-    }
-}
-
-impl Point {
-    /// Convert the point to a vector.
-    pub fn to_array(self) -> Vec2d {
-        [self.x, self.y]
-    }
-}
 
 /// A rectangle.
 #[derive(Clone, Copy, Debug)]
